Condense points for top topics
Previously similar points where listed under top topics, they were not merged. Here I also fix an issue with the condense_group logic that would loose a point when it clould only be paired with an already used point. For now these remainders are just added to the end of the list new group, even if they produce a better condensed point.

Note that the pairs are also sorted by the difference in their length, this means that what should be more similar pairs are condensed and spent first.