Implement formatted summary comparison view
As a final comparison, participants are going to compare the same
summary as the one they compared at at the second stage, only this time
with color and bold. This makes it a quick question to answer as they've
already seen the text and read it in the last question.