Require topic points to have variation
Now new points that are added to the list of topic points in the top
topic section must introduce a at least on new frame component. This
also involved flagging i and I as PERSON.

This will also make topic points respect points that have already been
listed in earlier sections.

Things was also rejected as an allowed topic word because it's clearly a
crappy topic word.