Update readme, remove instructions
The app is going to be a web app, not a series of cmd line tools. The recent edits to the compose file mean that the pipeline will no longer function in the same way. The analysis api and summarizer etc might as well be removed but they're being kept to reuse the code at a later date (counterpoints etc)