Include point_extractor draft
This is to eventually replace the points_api and be largely self
contained.

The matching is going to be written using the graph_match library,
rather than neo4j (https://github.com/charlieegan3/graph_match). This
will allow sentence graphs to be matched without persisting them to a
datastore in the process.