Implement HTTP post to corenlp
This is where the graph is going to come from and so this step is
required to move on. This is going to talk to corenlp alone in place of
the points_api.

In the future, ideally, the dependency parser would be contained within
the same container to avoid the request response overhead.