Generate a simplified verb index
This is going to be used in the new points extractor. Many matches are
made on the generic frame. This devalues the information in the frames
as it is often not used. I am still keen to use frames and know the
difference between them and the generic frame. However, parsing JSON is
easier in Rust (and indeed all compiled languages) if it is less
complex.