Add generic frame implementation
These are not quite as generic as I'd like, for a start there are two of
them. In the old cypher implementation I had optional matches, there are
not a thing in my graph query matching implementation.

This is based on the assumption that there are pretty much two classes,
np verb obj + np verb xcomp. These are both accounted for. Copulas are
not handled, I don't think they're needed here, the copula frames are
quite generic as it is. That is, they're just all added to all copula
verbs that come up (as per the Wikipedia list).